refactor(editor): Bind the injected upgrade-redirect guard via an app wrapper (no-changelog)
Per review, replace the bootstrap-registered module-global guard with an
explicit injected `{ guard }` parameter, wrapped by an app-side composable that
pre-binds `confirmIfBuilderStreaming`.

- The core (`usePageRedirectionHelper.core.ts`) now takes a required `{ guard }`
  and carries no feature dependency of its own — it is the unit that moves to
  `@n8n/composables` in N8N-71.
- The editor wrapper (`usePageRedirectionHelper.ts`) binds the builder-streaming
  guard, so every upgrade CTA stays covered by default with no call-site churn.
- Drops the module-level guard + register seam and the `init.ts` bootstrap
  wiring (both revert to untouched).

A `required` guard means a direct import of the core cannot silently skip the
confirmation — the caller must supply one. Behavior is unchanged.
